requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
Django==3.2.7
djangorestframework==3.12.4
idna==3.2
pycodestyle==2.7.0
pytz==2021.1
requests==2.26.0
selenium==3.141.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
webdriver-manager==3.4.2
